renamed to zed_camerainfo.py
